+++ RealmSwift/LinkingObjects.swift
@@ -136,7 +136,7 @@ public final class LinkingObjects<T: Object>: LinkingObjectsBase {
      */
     public func index(matching predicateFormat: String, _ args: Any...) -> Int? {
         return notFoundToNil(index: rlmResults.indexOfObject(with: NSPredicate(format: predicateFormat,
-                                                                               argumentArray: args)))
+                                                                               argumentArray: unwrapOptionals(in: args))))
     }
 
     // MARK: Object Retrieval
@@ -200,7 +200,8 @@ public final class LinkingObjects<T: Object>: LinkingObjectsBase {
      - parameter predicateFormat: A predicate format string, optionally followed by a variable number of arguments.
      */
     public func filter(_ predicateFormat: String, _ args: Any...) -> Results<T> {
-        return Results<T>(rlmResults.objects(with: NSPredicate(format: predicateFormat, argumentArray: args)))
+        return Results<T>(rlmResults.objects(with: NSPredicate(format: predicateFormat,
+                                                               argumentArray: unwrapOptionals(in: args))))
     }
 
     /**